Remove db_stress changes for now
